# backend-foryou/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ForYou API Documentation",
        "description": "API documentation for the ForYou project",
        "contact": {
            "email": "admin@foryou.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/proyectos": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Get list of proyectos",
                "description": "Returns list of proyectos",
                "operationId": "getProyectos",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Proyecto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Create a new proyecto",
                "description": "Creates a new proyecto",
                "operationId": "createProyecto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProyectoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proyecto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/proyectos/{id}": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Get proyecto by ID",
                "description": "Returns a single proyecto",
                "operationId": "getProyectoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of proyecto to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proyecto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Update an existing proyecto",
                "description": "Updates a proyecto",
                "operationId": "updateProyecto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of proyecto to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProyectoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proyecto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Delete a proyecto",
                "description": "Deletes a single proyecto",
                "operationId": "deleteProyecto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of proyecto to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Proyecto not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Proyecto": {
                "title": "Proyecto",
                "properties": {
                    "id_proyecto": {
                        "description": "ID del proyecto",
                        "type": "integer",
                        "format": "int64"
                    },
                    "observaciones": {
                        "description": "Observaciones del proyecto",
                        "type": "string"
                    },
                    "plano_url": {
                        "description": "URL del plano",
                        "type": "string",
                        "format": "url"
                    },
                    "plano_json": {
                        "description": "JSON del plano",
                        "type": "object"
                    },
                    "id_cliente": {
                        "description": "ID del cliente asociado",
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProyectoInput": {
                "title": "Proyecto Input",
                "properties": {
                    "observaciones": {
                        "description": "Observaciones del proyecto",
                        "type": "string"
                    },
                    "plano_url": {
                        "description": "URL del plano",
                        "type": "string",
                        "format": "url"
                    },
                    "plano_json": {
                        "description": "JSON del plano",
                        "type": "object"
                    },
                    "id_cliente": {
                        "description": "ID del cliente asociado",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}